+++ docs/changes.md
@@ -1,6 +1,9 @@
 #最新动态
 
-20201209
+2020-12-14
+  * 重新用object default实现decoder factory。
+
+2020-12-09
  * 将 audio\_decoder\_mp3 移动到 audio\_decoders 目录。
  * 将 audio\_decoder 和 audio\_decoder\_factory 移到 base 目录。
  * 将 audio\_device\_sdl 移到 audio\_device 目录。